+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file
 
+## Unreleased
+
+- Compatibility for Extraplanetary Launchpads
+- Compatibility for Kethane
+
+
 ## 0.0.1 - 2025-06-27
 
 - Add initial harmony patcher structure